+++ Cesium3DTilesSelection/src/RasterOverlayTileProvider.cpp
@@ -230,12 +230,14 @@ struct LoadResult {
  * @param pPrepareRendererResources The `IPrepareRendererResources`
  * @param pLogger The logger
  * @param loadedImage The `LoadedRasterOverlayImage`
+ * @param rendererOptions Renderer options
  * @return The `LoadResult`
  */
 static LoadResult createLoadResultFromLoadedImage(
     const std::shared_ptr<IPrepareRendererResources>& pPrepareRendererResources,
     const std::shared_ptr<spdlog::logger>& pLogger,
-    LoadedRasterOverlayImage&& loadedImage) {
+    LoadedRasterOverlayImage&& loadedImage,
+    const std::any& rendererOptions) {
   if (!loadedImage.image.has_value()) {
     SPDLOG_LOGGER_ERROR(
         pLogger,
@@ -271,8 +273,9 @@ static LoadResult createLoadResultFromLoadedImage(
 
     void* pRendererResources = nullptr;
     if (pPrepareRendererResources) {
-      pRendererResources =
-          pPrepareRendererResources->prepareRasterInLoadThread(image);
+      pRendererResources = pPrepareRendererResources->prepareRasterInLoadThread(
+          image,
+          rendererOptions);
     }
 
     LoadResult result;
@@ -311,12 +314,14 @@ void RasterOverlayTileProvider::doLoad(
   this->loadTileImage(tile)
       .thenInWorkerThread(
           [pPrepareRendererResources = this->getPrepareRendererResources(),
-           pLogger =
-               this->getLogger()](LoadedRasterOverlayImage&& loadedImage) {
+           pLogger = this->getLogger(),
+           rendererOptions = this->_pOwner->getOptions().rendererOptions](
+              LoadedRasterOverlayImage&& loadedImage) {
             return createLoadResultFromLoadedImage(
                 pPrepareRendererResources,
                 pLogger,
-                std::move(loadedImage));
+                std::move(loadedImage),
+                rendererOptions);
           })
       .thenInMainThread(
           [this, &tile, isThrottledLoad](LoadResult&& result) noexcept {